consolidate db.records module
