+++ mathesar/rpc/constraints.py
@@ -6,7 +6,7 @@ from typing import Optional, TypedDict, Union
 from modernrpc.core import rpc_method, REQUEST_KEY
 from modernrpc.auth.basic import http_basic_auth_login_required
 
-from db.constraints import(
+from db.constraints import (
     get_constraints_for_table,
     create_constraint,
     drop_constraint_via_oid,
