+++ mathesar/rpc/records.py
@@ -6,10 +6,14 @@ from typing import Any, Literal, Optional, TypedDict, Union
 from modernrpc.core import rpc_method, REQUEST_KEY
 from modernrpc.auth.basic import http_basic_auth_login_required
 
-from db.records.operations import delete as record_delete
-from db.records.operations import insert as record_insert
-from db.records.operations import select as record_select
-from db.records.operations import update as record_update
+from db.records import (
+    list_records_from_table,
+    get_record_from_table,
+    search_records_from_table,
+    delete_records_from_table,
+    add_record_to_table,
+    patch_record_in_table,
+)
 from mathesar.rpc.exceptions.handlers import handle_rpc_exceptions
 from mathesar.rpc.utils import connect
 
@@ -226,7 +230,7 @@ def list_(
     """
     user = kwargs.get(REQUEST_KEY).user
     with connect(database_id, user) as conn:
-        record_info = record_select.list_records_from_table(
+        record_info = list_records_from_table(
             conn,
             table_oid,
             limit=limit,
@@ -265,7 +269,7 @@ def get(
     """
     user = kwargs.get(REQUEST_KEY).user
     with connect(database_id, user) as conn:
-        record_info = record_select.get_record_from_table(
+        record_info = get_record_from_table(
             conn,
             record_id,
             table_oid,
@@ -307,7 +311,7 @@ def add(
     """
     user = kwargs.get(REQUEST_KEY).user
     with connect(database_id, user) as conn:
-        record_info = record_insert.add_record_to_table(
+        record_info = add_record_to_table(
             conn,
             record_def,
             table_oid,
@@ -350,7 +354,7 @@ def patch(
     """
     user = kwargs.get(REQUEST_KEY).user
     with connect(database_id, user) as conn:
-        record_info = record_update.patch_record_in_table(
+        record_info = patch_record_in_table(
             conn,
             record_def,
             record_id,
@@ -383,7 +387,7 @@ def delete(
     """
     user = kwargs.get(REQUEST_KEY).user
     with connect(database_id, user) as conn:
-        num_deleted = record_delete.delete_records_from_table(
+        num_deleted = delete_records_from_table(
             conn,
             record_ids,
             table_oid,
@@ -425,7 +429,7 @@ def search(
     """
     user = kwargs.get(REQUEST_KEY).user
     with connect(database_id, user) as conn:
-        record_info = record_select.search_records_from_table(
+        record_info = search_records_from_table(
             conn,
             table_oid,
             search=search_params,
