+++ mathesar/tests/rpc/test_records.py
@@ -56,7 +56,7 @@ def test_records_list(rf, monkeypatch):
         }
 
     monkeypatch.setattr(records, 'connect', mock_connect)
-    monkeypatch.setattr(records.record_select, 'list_records_from_table', mock_list_records)
+    monkeypatch.setattr(records, 'list_records_from_table', mock_list_records)
     expect_records_list = {
         "count": 50123,
         "results": [{"1": "abcde", "2": 12345}, {"1": "fghij", "2": 67890}],
@@ -116,7 +116,7 @@ def test_records_get(rf, monkeypatch):
         }
 
     monkeypatch.setattr(records, 'connect', mock_connect)
-    monkeypatch.setattr(records.record_select, 'get_record_from_table', mock_get_record)
+    monkeypatch.setattr(records, 'get_record_from_table', mock_get_record)
     expect_record = {
         "count": 1,
         "results": [{"1": "abcde", "2": 12345}, {"1": "fghij", "2": 67890}],
@@ -169,7 +169,7 @@ def test_records_add(rf, monkeypatch):
         }
 
     monkeypatch.setattr(records, 'connect', mock_connect)
-    monkeypatch.setattr(records.record_insert, 'add_record_to_table', mock_add_record)
+    monkeypatch.setattr(records, 'add_record_to_table', mock_add_record)
     expect_record = {
         "results": [record_def],
         "linked_record_summaries": {"2": {"12345": "blkjdfslkj"}},
@@ -226,7 +226,7 @@ def test_records_patch(rf, monkeypatch):
         }
 
     monkeypatch.setattr(records, 'connect', mock_connect)
-    monkeypatch.setattr(records.record_update, 'patch_record_in_table', mock_patch_record)
+    monkeypatch.setattr(records, 'patch_record_in_table', mock_patch_record)
     expect_record = {
         "results": [record_def | {"3": "another"}],
         "linked_record_summaries": {"2": {"12345": "blkjdfslkj"}},
@@ -272,7 +272,7 @@ def test_records_delete(rf, monkeypatch):
         return 2
 
     monkeypatch.setattr(records, 'connect', mock_connect)
-    monkeypatch.setattr(records.record_delete, 'delete_records_from_table', mock_delete_records)
+    monkeypatch.setattr(records, 'delete_records_from_table', mock_delete_records)
     expect_result = 2
     actual_result = records.delete(
         record_ids=record_ids, table_oid=table_oid, database_id=database_id, request=request
@@ -316,7 +316,7 @@ def test_records_search(rf, monkeypatch):
         }
 
     monkeypatch.setattr(records, 'connect', mock_connect)
-    monkeypatch.setattr(records.record_select, 'search_records_from_table', mock_search_records)
+    monkeypatch.setattr(records, 'search_records_from_table', mock_search_records)
     expect_records_list = {
         "count": 50123,
         "results": [{"1": "abcde", "2": 12345}, {"1": "fghij", "2": 67890}],
